requirements.txt
langchain
langchain_community
ipykernel
python-dotenv
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain-openai
langchain_chroma
beautifulsoup4
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
pydantic
langgraph